+++ gatling-grpc-kt/src/main/kotlin/com/github/phisgr/gatling/kt/grpc/action/StreamStartBuilder.kt
@@ -2,11 +2,10 @@
 
 package com.github.phisgr.gatling.kt.grpc.action
 
-import com.github.phisgr.gatling.generic.SessionCombiner
 import com.github.phisgr.gatling.grpc.check.GrpcResponse
 import com.github.phisgr.gatling.grpc.check.StreamCheck
 import com.github.phisgr.gatling.grpc.stream.EventExtractor
-import com.github.phisgr.gatling.grpc.stream.TimestampExtractor
+import com.github.phisgr.gatling.kt.generic.SessionCombiner
 import com.github.phisgr.gatling.kt.grpc.StreamEndLog
 import com.github.phisgr.gatling.kt.grpc.internal.GrpcStreamEnd
 import com.github.phisgr.gatling.kt.grpc.internal.buildStream
@@ -14,6 +13,7 @@ import com.github.phisgr.gatling.kt.grpc.internal.buildStreamEnd
 import com.github.phisgr.gatling.kt.grpc.request.CallDefinition
 import com.github.phisgr.gatling.kt.grpc.statusCode
 import com.github.phisgr.gatling.kt.grpc.statusDescription
+import com.github.phisgr.gatling.kt.grpc.stream.TimestampExtractor
 import com.github.phisgr.gatling.kt.grpc.trailer
 import com.github.phisgr.gatling.kt.internal.*
 import io.gatling.commons.validation.Validation
@@ -26,7 +26,9 @@ import scala.collection.immutable.Seq
 import scala.runtime.`Null$`
 import java.util.function.BiPredicate
 import java.util.function.Predicate
+import com.github.phisgr.gatling.generic.SessionCombiner as SessionCombinerS
 import com.github.phisgr.gatling.grpc.action.StreamStartBuilder as StreamStartBuilderS
+import com.github.phisgr.gatling.grpc.stream.TimestampExtractor as TimestampExtractorS
 import io.gatling.core.session.Session as SessionS
 
 /**
@@ -82,6 +84,10 @@ abstract class StreamStartBuilder<Self : StreamStartBuilder<Self, Req, Res, Wrap
     fun timestampExtractor(extractor: TimestampExtractor<Res>): Self =
         wrap(asScala().timestampExtractor(extractor))
 
+    @Deprecated("Use the overload for the Kotlin wrapper type.", level = DeprecationLevel.HIDDEN)
+    fun timestampExtractor(extractor: TimestampExtractorS<Res>): Self =
+        wrap(asScala().timestampExtractor(extractor))
+
     /**
      * This exposes the internal StatsEngine of Gatling.
      * No wrapper class around them, as this is a niche, advanced API.
@@ -90,6 +96,10 @@ abstract class StreamStartBuilder<Self : StreamStartBuilder<Self, Req, Res, Wrap
         wrap(asScala().eventExtractor(extractor))
 
     fun sessionCombiner(sessionCombiner: SessionCombiner): Self =
+        wrap(asScala().sessionCombiner(sessionCombiner))
+
+    @Deprecated("Use the overload for the Kotlin wrapper type.", level = DeprecationLevel.HIDDEN)
+    fun sessionCombiner(sessionCombiner: SessionCombinerS): Self =
         wrap(asScala().sessionCombiner(sessionCombiner))