+++ src/main/scala/com/github/phisgr/gatling/grpc/action/StreamCancelBuilder.scala
@@ -8,17 +8,15 @@ import io.gatling.core.action.builder.ActionBuilder
 import io.gatling.core.session.{Expression, Session}
 import io.gatling.core.structure.ScenarioContext
 
-class StreamCancelBuilder(requestName: Expression[String], streamName: String, direction: String) extends ActionBuilder {
+class StreamCancelBuilder(requestName: Expression[String], streamName: Expression[String], direction: String) extends ActionBuilder {
   override def build(ctx: ScenarioContext, next: Action): Action =
-    new StreamMessageAction(requestName, ctx, next, baseName = "StreamClose", direction) {
-      // close over the string rather than the outer class
-      private[this] val streamName = StreamCancelBuilder.this.streamName
+    new StreamMessageAction(requestName, streamName, ctx, next, baseName = "StreamClose", direction) {
       override def sendRequest(session: Session): Validation[Unit] = forToMatch {
         for {
-          call <- fetchCall[Cancellable](streamName, session)
+          call <- fetchCall(classOf[Cancellable], session)
         } yield {
-          logger.debug(s"Cancelling $direction stream '$streamName': Scenario '${session.scenario}', UserId #${session.userId}")
-          call.cancel(session, next)
+          logger.debug(s"Cancelling ${this.direction} stream '${call.streamName}': Scenario '${session.scenario}', UserId #${session.userId}")
+          call.cancel(session, this.next)
         }
       }
     }